RTPsender update

+++ source/codestream/j2kbuf.cpp
@@ -24,7 +24,7 @@ void J2kBuf::reset(uint8_t* const in) {
 }
 
 uint8_t J2kBuf::get_bit() {
-    if (bit_pos & 0x80) {
+    if (bit_pos & static_cast<uint8_t>(0x80)) {
         termination_check();
         bit_purge = buf_ptr[byte_pos];
     }
@@ -33,7 +33,7 @@ uint8_t J2kBuf::get_bit() {
     if (bit_pos & static_cast<uint8_t>(0x01)) {
         bit_pos = static_cast<uint8_t>(0x80);
         // if (buf_ptr[byte_pos] == 0xFF) { // bit stuffing
-        if (bit_purge == 0xFF) { // bit stuffing
+        if (bit_purge == static_cast<uint8_t>(0xFF)) { // bit stuffing
             bit_pos >>= 1;
         }
         advance_byte_pos(1);
